Fileset population with 0% pre-allocation could fail
Fileset population with 0% pre-allocation could fail if a large number
of files were defined (e.g., 2,000,0000). In this case randno could
equal to RAND_MAX for some file and Filebench tried to create a file
that it was not supposed to. Then it was failing with the following
error:

Failed to pre-allocate file /test/root/deepfileset/00000001/00372688:
No such file or directory on line 17

Fixed by this patch.