"Error on line X" changed to "Error around line X"
The rule's handler is called only when bison detects the complete rule.
To detect the rule it often needs to reach to the next token that might be
several (usually empty) lines below. As a result the error
might be not exactly on the line where lex currently is, but earlier.