Webserver personality: making fileset readonly
Opening files in read/write mode can cause problems;
for example under NFSv4 it automatically tries to lock
a file or create a delegation.  Some other filesystems
like unionfs, when you open a file in write mode,
they do a copyup on it so if the file does not
have to be open in write mode it should not be.

We assume that modern  Web-servers are smart enough
to open static files in read-only mode.