Making mt_* and mts_* functions non-inline
Depending on the compiler specifics, "inline" keyword can be treated very
differently. Mtwist library used several macroses to define the behavior (8
different states!), but still was not covering non-gcc compilers reliably.
Futhermore, after the release of gcc version 5, which changed default inline
treatment, the macroses stopped working again and compilation broke.

Since FB should work on many platforms with very different compilers, this
commit removes the inline definitions of mt_* and mts_* functions and moves
them to the *.c file from the header. This makes raw random number generation a
bit slower but given the weight of complete Filebench execution it is a minor
degradation. In exchange we get portability.