Changing the way fds are picked when "opennext" is specified
Previously, for unknown reasons, the selection of
fd was tied to the size of the fileset. E.g., if the
fileset size was larger than 128 files, then "opennext"
was failing, because 128 / 2 = 64 = MAXIMUM_NUMBER_OF_FDS.
This was a weird logic and replaced by a new more clear one.

All flowops with opennext attribute cycle through
0 to MAXIMUM_NUMBER_OF_FD (128) descriptors (backwards).
Notice, that the the cycle counter is per-thread (not per-flowop)