Removing redundant caret (^) symbols from the whitestring definition
Either one who wrote the code assumed incorrectly that every excluded
character in square brackets has to be preceeded with ^ in bison, or it
used to be the case for some old version of bison.  But in modern times,
bison only wants one caret following the opening bracket. Otherwise
it exludes the caret itself from the string.

Fixes issue #86