check fs_readonly flag in fileset_alloc_file
If one has preallocated a fileset and is then attempting to perform read-only operations on it (e.g. through a read-only mount) this will fail because fileset_alloc_file will attempt to open the file O_RDWR. Check the fs_readonly fileset flag and open the given path with O_RDONLY if user specifies fileset as "readonly".